refactor(execution): co-locate per-intent tests with their code
Follow-up to the previous execution split: tests stayed in mod.rs
when the code moved to per-intent files, leaving mod.rs at 74 LOC of
code and 938 LOC of tests testing behaviour that lives elsewhere.

Each Intent's tests now live next to its impl:

- link.rs   — Link-intent tests (deploy + conflict + force +
              ancestor-cycle suite)
- stage.rs  — Stage-intent tests (data-link, path auto-chmod,
              shell handler chmod-skip, dry-run report)
- run.rs    — Run-intent tests (sentinel record / skip /
              provision_rerun)
- mod.rs    — Cross-intent dispatcher tests only:
              dry_run_does_not_modify_filesystem and
              execute_multiple_intents_sequentially
- test_support.rs — Shared MockCommandRunner + make_datastore,
                    pub(super) so siblings can reach via
                    super::test_support::*

No production change; same 1140 tests still pass.
